+++ src/ViewInstance.js
@@ -9,7 +9,6 @@ import {
 import {
   flowRight,
   isEmpty,
-  pick,
 } from 'lodash';
 
 import {
@@ -38,9 +37,10 @@ import {
   handleKeyCommand,
   isInstanceShadowCopy,
   isMARCSource,
+  getLinkedAuthorityIds,
 } from './utils';
 import {
-  AUTHORITY_LINKED_FIELDS,
+  CONSORTIUM_PREFIX,
   HTTP_RESPONSE_STATUS_CODES,
   indentifierTypeNames,
   INSTANCE_SHARING_STATUSES,
@@ -170,6 +170,12 @@ class ViewInstance extends React.Component {
       accumulate: true,
       throwErrors: false,
     },
+    authorities: {
+      type: 'okapi',
+      path: 'authority-storage/authorities',
+      accumulate: true,
+      throwErrors: false,
+    }
   });
 
   constructor(props) {
@@ -521,35 +527,32 @@ class ViewInstance extends React.Component {
       });
   }
 
-  checkIfHasLinkedAuthorities = (instance = {}) => {
-    const linkedAuthorities = pick(this.props.selectedInstance, AUTHORITY_LINKED_FIELDS);
+  checkIfHasLinkedAuthorities = () => {
+    const { selectedInstance } = this.props;
+    const authorityIds = getLinkedAuthorityIds(selectedInstance);
 
-    const findLinkedAuthorities = authorities => {
-      const linkedAuthoritiesLength = AUTHORITY_LINKED_FIELDS.reduce((total, field) => {
-        const authoritiesAmount = authorities[field].filter(item => item.authorityId).length;
-        return total + authoritiesAmount;
-      }, 0);
-
-      return {
-        hasLinkedAuthorities: !!linkedAuthoritiesLength,
-        linkedAuthoritiesLength,
-      };
-    };
-
-    const {
-      hasLinkedAuthorities,
-      linkedAuthoritiesLength,
-    } = findLinkedAuthorities(linkedAuthorities);
-
-    if (hasLinkedAuthorities) {
-      this.setState({
-        linkedAuthoritiesLength,
-        isShareLocalInstanceModalOpen: false,
-        isUnlinkAuthoritiesModalOpen: true,
-      });
-    } else {
-      this.handleShareLocalInstance(instance);
+    if (isEmpty(authorityIds)) {
+      this.handleShareLocalInstance(selectedInstance);
+      return;
     }
+
+    this.props.mutator.authorities.GET({
+      params: {
+        query: `id==(${authorityIds.join(' or ')})`,
+      }
+    }).then(({ authorities }) => {
+      const localAuthorities = authorities.filter(authority => !authority.source.startsWith(CONSORTIUM_PREFIX));
+
+      if (localAuthorities.length) {
+        this.setState({
+          linkedAuthoritiesLength: localAuthorities.length,
+          isShareLocalInstanceModalOpen: false,
+          isUnlinkAuthoritiesModalOpen: true,
+        });
+      } else {
+        this.handleShareLocalInstance(selectedInstance);
+      }
+    });
   }
 
   toggleCopyrightModal = () => {
@@ -1031,7 +1034,7 @@ class ViewInstance extends React.Component {
               message={<FormattedMessage id="ui-inventory.shareLocalInstance.modal.message" values={{ instanceTitle: instance?.title }} />}
               confirmLabel={<FormattedMessage id="ui-inventory.shareLocalInstance.modal.confirmButton" />}
               onCancel={() => this.setState({ isShareLocalInstanceModalOpen: false })}
-              onConfirm={() => this.checkIfHasLinkedAuthorities(instance)}
+              onConfirm={this.checkIfHasLinkedAuthorities}
             />
 
             <ConfirmationModal
@@ -1094,6 +1097,9 @@ ViewInstance.propTypes = {
     }).isRequired,
     shareInstance: PropTypes.shape({
       POST: PropTypes.func.isRequired,
+      GET: PropTypes.func.isRequired,
+    }).isRequired,
+    authorities: PropTypes.shape({
       GET: PropTypes.func.isRequired,
     }).isRequired,
   }),