+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [10.0.3] IN PROGRESS
 
 * ECS: Show info message when user in member tenant tries to view shared instance details without permission. Refs UIIN-2590.
+* Show only local MARC Authorities when share local instance. Fixes UIIN-2647.
 
 ## [10.0.2](https://github.com/folio-org/ui-inventory/tree/v10.0.2) (2023-11-07)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v10.0.1...v10.0.2)